source/requirements.txt
apache-airflow==2.2.3
apache-airflow[postgres]==2.2.3
psycopg2-binary==2.9.1
markupsafe==2.0.1
pyarrow==5.0.0
pandas==1.3.4
